+++ outputs.tf
@@ -17,3 +17,13 @@ output "frontdoor_frontend_endpoints" {
   description = "The IDs of the frontend endpoints."
   value       = azurerm_frontdoor.frontdoor.frontend_endpoints
 }
+
+output "frontdoor_address_prefixes_ipv4" {
+  description = "IPv4 address ranges used by the FrontDoor service"
+  value       = [for ip in jsondecode(data.external.frontdoor_ips.result.addressPrefixes) : ip if length(regexall("\\.", ip)) > 0]
+}
+
+output "frontdoor_address_prefixes_ipv6" {
+  description = "IPv6 address ranges used by the FrontDoor service"
+  value       = [for ip in jsondecode(data.external.frontdoor_ips.result.addressPrefixes) : ip if length(regexall(":", ip)) > 0]
+}