+++ versions.tf
@@ -5,5 +5,9 @@ terraform {
       source  = "hashicorp/azurerm"
       version = ">= 2.60"
     }
+    external = {
+      source  = "hashicorp/external"
+      version = ">= 2"
+    }
   }
 }